+++ examples/js/nextjs/pages/borked.js
@@ -1,4 +1,0 @@
-export default () =>
-  <div>
-    <p>{this.stuff.uhoh}</p>
-  </div>