+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v7.17.1 (2022-06-29)
+
+### Fixed
+
+- (react-native) Update bugsnag-android from v5.22.4 to [v5.23.0](https://github.com/bugsnag/bugsnag-android/blob/master/CHANGELOG.md#5230-2022-06-20)
+- (react-native) Update bugsnag-cocoa from v6.18.0 to [v6.18.1](https://github.com/bugsnag/bugsnag-cocoa/blob/master/CHANGELOG.md#6181-2022-06-22)
+- (react-native) Add support for React Native v0.69 [#1775](https://github.com/bugsnag/bugsnag-js/pull/1775)
+- (react-native) Add 0.69 test fixtures and CI coverage [#1775](https://github.com/bugsnag/bugsnag-js/pull/1775)
+- (plugin-react-native-unhandled-rejection) Fix TypeError when using react native 0.63 or 0.64 [#1770](https://github.com/bugsnag/bugsnag-js/pull/1770)
+
 ## v7.17.0 (2022-06-20)
 
 ### Added